% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGD_Adam.R
\name{NR}
\alias{NR}
\title{This is Newton-Raphson}
\usage{
NR(delta, z, knot = 10, facility = NULL, theta_NR = NULL, M_stop = 1000)

}
\arguments{
\item{delta}{event indicator}

\item{z}{Covariate matrix}

\item{knot}{number of basis functions for time-varing effects. the default value is 10.}

\item{facility}{strata.}

\item{theta_NR}{the initial value}

\item{M_stop}{maximum stopping iterations}

}
\description{
This is SGD-Adam version.
}
\examples{
SGD_Adam(delta, z, time, knot = 10, facility = NULL, theta_SGD = NULL, 
    M_stop = 10000, tol = 10^(-6)) 
}
\import{splines}