% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGD.R
\name{SGD}
\alias{SGD}
\title{This is the basic SGD version.}
\usage{
SGD(delta, z, time, knot = 10, facility = NULL, theta_SGD = NULL,
  M_stop = 10000, tol = 10^(-6))
}
\arguments{
\item{delta}{event indicator}

\item{z}{Covariate matrix}

\item{time}{observed event time}

\item{knot}{number of basis functions for time-varing effects. the default value is 10.}

\item{facility}{strata.}

\item{M_stop}{maximum stopping iterations}

\item{tol}{the convergence threshold}
}
\description{
This is the basic SGD version.
}
\examples{
SGD(delta, z, time, knot = 10, facility = NULL, theta_SGD = NULL, 
    M_stop = 10000, tol = 10^(-6)) 
}
\import{splines}