\name{SurvTime-package}
\alias{SurvTime-package}
\alias{SurvTime}
\docType{package}
\title{
\packageTitle{SurvTime}
}
\description{
\packageDescription{SurvTime}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SurvTime}
\packageIndices{SurvTime}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SurvTime}

Maintainer: \packageMaintainer{SurvTime}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
